+++ Dto/DistributorReportDto.cs
@@ -3,6 +3,6 @@ public class DistributorAccountBalanceDto
     public string DistributorId { get; set; } = "";
     public string DistributorName { get; set; } = "";
     public string DistributorEmail { get; set; } = "";
-    public double TotalWallet { get; set; }
-    public double CommissionAmount { get; set; }
+    public decimal TotalWallet { get; set; }
+    public decimal CommissionAmount { get; set; }
 }
